+++ Source/TimeWarp.State.Plus/Features/ActionTracking/Actions/ActionTrackingState.CompleteProcessing.cs
@@ -1,6 +1,6 @@
-﻿namespace TimeWarp.Features.Processing;
+﻿namespace TimeWarp.Features.ActionTracking;
 
-public partial class ActiveActionState
+public partial class ActionTrackingState
 {
   public static class CompleteProcessing
   {
@@ -12,11 +12,11 @@ public partial class ActiveActionState
       IStore store
     ): ActionHandler<Action>(store)
     {
-      private ActiveActionState ActiveActionState => Store.GetState<ActiveActionState>();
+      private ActionTrackingState ActionTrackingState => Store.GetState<ActionTrackingState>();
       
       public override Task Handle(Action action, CancellationToken cancellationToken)
       {
-        ActiveActionState.ActiveActionsList.Remove(action.TheAction);
+        ActionTrackingState.ActiveActionsList.Remove(action.TheAction);
         return Task.CompletedTask;
       }
     }
