+++ Source/TimeWarp.State.Plus/Features/ActionTracking/Pipeline/TrackActionAttribute.cs
@@ -1,4 +1,4 @@
-﻿namespace TimeWarp.Features.Processing;
+﻿namespace TimeWarp.Features.ActionTracking;
 
 [AttributeUsage(AttributeTargets.Class, AllowMultiple = false)]
 public class TrackActionAttribute : Attribute { }
